Bugfixes, add LdapLoginModule + performance improvements for session beans
