+++ UPGRADE-1.1.4.md
@@ -1,0 +1,3 @@
+# Upgrade from 1.1.3 to 1.1.4
+
+Updating from 1.1.3 to 1.1.4 doesn't have any impacts. Please read the apropriate UPGRADE-1.x.x files for updates from older versions to 1.1.3.
