+++ src/AppserverIo/Appserver/Core/AbstractManager.php
@@ -27,7 +27,6 @@ use AppserverIo\Psr\Application\ManagerInterface;
 use AppserverIo\Psr\Application\ApplicationInterface;
 use AppserverIo\Psr\Application\ManagerConfigurationInterface;
 use AppserverIo\Psr\Naming\InitialContext as NamingDirectory;
-use AppserverIo\Appserver\ServletEngine\RequestHandler;
 
 /**
  * Abstract manager implementation.
@@ -184,30 +183,6 @@ abstract class AbstractManager extends GenericStackable implements ManagerInterf
     public function newReflectionClass($className)
     {
         return $this->getApplication()->search('ProviderInterface')->newReflectionClass($className);
-    }
-
-    /**
-     * This returns a proxy to the requested session bean.
-     *
-     * @param string $lookupName The lookup name for the requested session bean
-     * @param string $sessionId  The session-ID if available
-     *
-     * @return \AppserverIo\RemoteMethodInvocation\RemoteObjectInterface The proxy instance
-     */
-    public function lookupProxy($lookupName, $sessionId = null)
-    {
-
-        // load the initial context instance
-        $initialContext = $this->getInitialContext();
-
-        // query whether a request context is available
-        if ($servletRequest = RequestHandler::getRequestContext()) {
-            // inject the servlet request to handle SFSBs correctly
-            $initialContext->injectServletRequest($servletRequest);
-        }
-
-        // lookup the proxy by the name and session ID if available
-        return $initialContext->lookup($lookupName, $sessionId);
     }
 
     /**
