+++ src/AppserverIo/Appserver/ServletEngine/Session/AbstractSessionHandler.php
@@ -20,7 +20,6 @@
 
 namespace AppserverIo\Appserver\ServletEngine\Session;
 
-use AppserverIo\Lang\Reflection\ReflectionClass;
 use AppserverIo\Psr\Servlet\ServletSessionInterface;
 use AppserverIo\Appserver\ServletEngine\Http\Session;
 use AppserverIo\Appserver\ServletEngine\SessionSettingsInterface;
