+++ src/AppserverIo/Appserver/ServletEngine/Session/SessionHandlerFactory.php
@@ -22,7 +22,6 @@ namespace AppserverIo\Appserver\ServletEngine\Session;
 
 use AppserverIo\Lang\Reflection\ReflectionClass;
 use AppserverIo\Appserver\ServletEngine\SessionSettingsInterface;
-use AppserverIo\Appserver\ServletEngine\SessionMarshallerInterface;
 use AppserverIo\Appserver\Core\Api\Node\SessionHandlerNodeInterface;
 
 /**
